+++ intelmq/lib/bot.py
@@ -41,6 +41,8 @@ class Bot(object):
     module = None
     name = None
 
+    _message_processed_verb = 'Processed'
+
     def __init__(self, bot_id: str):
         self.__log_buffer = []
         self.parameters = Parameters()
@@ -285,7 +287,9 @@ class Bot(object):
             self.logger.exception('Error during shutdown of bot.')
 
         if self.__message_counter:
-            self.logger.info("Processed %d messages since last logging.", self.__message_counter)
+            self.logger.info("%s %d messages since last logging.",
+                             self._message_processed_verb,
+                             self.__message_counter)
 
         self.__disconnect_pipelines()